last changes

+++ FirstProject/Controllers/UsersController.cs
@@ -1,4 +1,5 @@
 ﻿using Data;
+using FirstProject.Business;
 using FirstProject.ViewModels;
 using System;
 using System.Data;
@@ -12,20 +13,35 @@ namespace FirstProject.Controllers
     public class UsersController : Controller
     {
         private FirstProjectEntities db = new FirstProjectEntities();
+        private LoginBusiness loginRestriction = new LoginBusiness();
 
         // GET: Users
         
         public ActionResult Index()
         {
-            try
+            
+            if (loginRestriction.IsRestricted((int?)Session["agencyId"]) == true)
             {
-                db.Configuration.LazyLoadingEnabled = false;
-                return View(db.Users.ToList());
+                return RedirectToAction("Login", "Users");
             }
-            catch(Exception ex)
+            else
             {
-                return null;
-            };
+
+
+                try
+                {
+                    db.Configuration.LazyLoadingEnabled = false;
+                    return View(db.Users.ToList());
+                }
+                catch (Exception ex)
+                {
+                    return null;
+                };
+
+                
+            }
+            
+            
         }
     
         
@@ -33,21 +49,29 @@ namespace FirstProject.Controllers
         // GET: Users/Details/5
         public ActionResult Details(int? id)
         {
-            if (id == null)
+            if (loginRestriction.IsRestricted((int?)Session["agencyId"]) == true)
             {
-                return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
+                return RedirectToAction("Login", "Users");
             }
-            User user = db.Users.Find(id);
-            if (user == null)
+            else
             {
-                return HttpNotFound();
+                if (id == null)
+                {
+                    return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
+                }
+                User user = db.Users.Find(id);
+                if (user == null)
+                {
+                    return HttpNotFound();
+                }
+                return View(user);
             }
-            return View(user);
         }
 
         // GET: Users/Create
         public ActionResult Create()
         {
+
             return View();
         }
 
@@ -66,16 +90,23 @@ namespace FirstProject.Controllers
         // GET: Users/Edit/5
         public ActionResult Edit(int? id)
         {
-            if (id == null)
+            if (loginRestriction.IsRestricted((int?)Session["agencyId"]) == true)
             {
-                return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
+                return RedirectToAction("Login", "Users");
             }
-            User user = db.Users.Find(id);
-            if (user == null)
+            else
             {
-                return HttpNotFound();
+                if (id == null)
+                {
+                    return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
+                }
+                User user = db.Users.Find(id);
+                if (user == null)
+                {
+                    return HttpNotFound();
+                }
+                return View(user);
             }
-            return View(user);
         }
 
         // POST: Users/Edit/5
@@ -85,28 +116,42 @@ namespace FirstProject.Controllers
         [ValidateAntiForgeryToken]
         public ActionResult Edit([Bind(Include = "Id,Username,Password,IsActive,LastLoginDate")] User user)
         {
-            if (ModelState.IsValid)
+            if (loginRestriction.IsRestricted((int)Session["agencyId"]) == true)
             {
-                db.Entry(user).State = EntityState.Modified;
-                db.SaveChanges();
-                return RedirectToAction("Index");
+                return RedirectToAction("Login", "Users");
             }
-            return View(user);
+            else
+            {
+                if (ModelState.IsValid)
+                {
+                    db.Entry(user).State = EntityState.Modified;
+                    db.SaveChanges();
+                    return RedirectToAction("Index");
+                }
+                return View(user);
+            }
         }
 
         // GET: Users/Delete/5
         public ActionResult Delete(int? id)
         {
-            if (id == null)
+            if (loginRestriction.IsRestricted((int?)Session["agencyId"]) == true)
             {
-                return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
+                return RedirectToAction("Login", "Users");
             }
-            User user = db.Users.Find(id);
-            if (user == null)
+            else
             {
-                return HttpNotFound();
+                if (id == null)
+                {
+                    return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
+                }
+                User user = db.Users.Find(id);
+                if (user == null)
+                {
+                    return HttpNotFound();
+                }
+                return View(user);
             }
-            return View(user);
         }
 
         // POST: Users/Delete/5
@@ -114,10 +159,17 @@ namespace FirstProject.Controllers
         [ValidateAntiForgeryToken]
         public ActionResult DeleteConfirmed(int id)
         {
-            User user = db.Users.Find(id);
-            db.Users.Remove(user);
-            db.SaveChanges();
-            return RedirectToAction("Index");
+            if (loginRestriction.IsRestricted((int?)Session["agencyId"]) == true)
+            {
+                return RedirectToAction("Login", "Users");
+            }
+            else
+            {
+                User user = db.Users.Find(id);
+                db.Users.Remove(user);
+                db.SaveChanges();
+                return RedirectToAction("Index");
+            }
         }
 
         protected override void Dispose(bool disposing)
